Admin mora da promijeni lozniku

+++ Back-end/src/main/java/mrsisa/project/controller/AdminController.java
@@ -2,7 +2,7 @@ package mrsisa.project.controller;
 
 import mrsisa.project.dto.*;
 import mrsisa.project.model.Administrator;
-import mrsisa.project.model.Instructor;
+import mrsisa.project.model.LoyaltyProgram;
 import mrsisa.project.model.Payment;
 import mrsisa.project.service.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +25,7 @@ public class AdminController {
 
     @GetMapping(value = "/get")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<AdminDTO> getInstructor(Principal userP){
+    public ResponseEntity<AdminDTO> getAdmin(Principal userP){
         return new ResponseEntity<>(new AdminDTO(adminService.findAdminByUsername(userP.getName())), HttpStatus.OK);
     }
 
@@ -42,8 +42,11 @@ public class AdminController {
 
     @GetMapping(path = "/getProfileDeletionRequests")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<List<ProfileDeletionReasonDTO>> getProfileDeletionRequests()
+    public ResponseEntity<List<ProfileDeletionReasonDTO>> getProfileDeletionRequests(Principal userP)
     {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         List<ProfileDeletionReasonDTO> list = adminService.getProfileDeletionReasons();
         return new ResponseEntity<>(list, HttpStatus.OK);
     }
@@ -53,8 +56,12 @@ public class AdminController {
     public ResponseEntity<String> confirmDeletion(@PathVariable("userId") Long userId,
                                                   @PathVariable("pdrId") Long pdrId,
                                                   @PathVariable("confirmation") String confirmation,
-                                                  @RequestPart("message") String message)
+                                                  @RequestPart("message") String message,
+                                                  Principal userP)
     {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         boolean delete = Boolean.parseBoolean(confirmation);
         boolean success = adminService.deleteAccount(userId, pdrId, delete, message);
         if (!success)
@@ -66,8 +73,11 @@ public class AdminController {
 
     @GetMapping(path = "/getUserRegistrationRequests")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<List<RegistrationRequestDTO>> getUserRegistrationRequests()
+    public ResponseEntity<List<RegistrationRequestDTO>> getUserRegistrationRequests(Principal userP)
     {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         List<RegistrationRequestDTO> list = adminService.getRegistrationRequests();
         return new ResponseEntity<>(list, HttpStatus.OK);
     }
@@ -77,8 +87,12 @@ public class AdminController {
     public ResponseEntity<String> registerUser(@PathVariable("userId") Long userId,
                                                @PathVariable("regId") Long regId,
                                                @PathVariable("confirmation") String confirmation,
-                                               @RequestPart("message") String message)
+                                               @RequestPart("message") String message,
+                                               Principal userP)
     {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         boolean register = Boolean.parseBoolean(confirmation);
         boolean success = adminService.registerUser(userId, regId, register, message);
         if (!success)
@@ -91,7 +105,12 @@ public class AdminController {
     @GetMapping(path = "/getChartData/{startDate}&{endDate}")
     @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<ChartDataDTO>> getChartData(@PathVariable String startDate,
-                                                           @PathVariable String endDate) {
+                                                           @PathVariable String endDate,
+                                                           Principal userP)
+    {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         List<ChartDataDTO> chartData = adminService.getChartData(startDate, endDate);
         if (chartData.size() == 0)
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(chartData);
@@ -100,16 +119,42 @@ public class AdminController {
 
     @GetMapping(path = "/getPaymentConfig")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Payment> getPaymentConfig() {
+    public ResponseEntity<Payment> getPaymentConfig(Principal userP) {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         Payment payment = adminService.getPaymentConfig();
         return new ResponseEntity<>(payment, HttpStatus.OK);
     }
 
     @PutMapping(path = "/updatePaymentConfig")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<String> updatePaymentConfig(@RequestBody Payment payment) {
+    public ResponseEntity<String> updatePaymentConfig(@RequestBody Payment payment, Principal userP) {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
         adminService.updatePaymentConfig(payment);
         return ResponseEntity.status(HttpStatus.OK).body("Payment config successfully updated!");
+    }
+
+    @GetMapping(path = "/getLoyaltyProgram")
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<LoyaltyProgram> getLoyaltyProgram(Principal userP) {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        LoyaltyProgram loyaltyProgram = adminService.getLoyaltyProgram();
+        return new ResponseEntity<>(loyaltyProgram, HttpStatus.OK);
+    }
+
+    @PutMapping(path = "/updateLoyaltyProgram")
+    @PreAuthorize("hasRole('ADMIN')")
+    public ResponseEntity<String> updateLoyaltyProgram(@RequestBody LoyaltyProgram loyaltyProgram, Principal userP) {
+        Administrator admin = adminService.findAdminByUsername(userP.getName());
+        if (admin.getLastPasswordResetDate() == null)
+            return new ResponseEntity<>(null, HttpStatus.FORBIDDEN);
+        adminService.updateLoyaltyProgram(loyaltyProgram);
+        return ResponseEntity.status(HttpStatus.OK).body("Loyalty program successfully updated!");
     }
